+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/statictools/PiSDFFlattenerTask.java
@@ -36,17 +36,15 @@
 /**
  *
  */
-package org.preesm.algorithm.pisdf.pimm2flat;
+package org.preesm.model.pisdf.statictools;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import org.eclipse.core.runtime.IProgressMonitor;
-import org.preesm.algorithm.model.visitors.VisitorOutput;
 import org.preesm.commons.logger.PreesmLogger;
 import org.preesm.model.pisdf.PiGraph;
-import org.preesm.model.scenario.PreesmScenario;
 import org.preesm.workflow.elements.Workflow;
 import org.preesm.workflow.implement.AbstractTaskImplementation;
 import org.preesm.workflow.implement.AbstractWorkflowNodeImplementation;
@@ -55,7 +53,9 @@ import org.preesm.workflow.implement.AbstractWorkflowNodeImplementation;
  * @author farresti
  *
  */
-public class StaticPiMM2FlatPiMMTask extends AbstractTaskImplementation {
+public class PiSDFFlattenerTask extends AbstractTaskImplementation {
+
+  final Logger logger = PreesmLogger.getLogger();
 
   /*
    * (non-Javadoc)
@@ -66,17 +66,10 @@ public class StaticPiMM2FlatPiMMTask extends AbstractTaskImplementation {
   @Override
   public Map<String, Object> execute(final Map<String, Object> inputs, final Map<String, String> parameters,
       final IProgressMonitor monitor, final String nodeName, final Workflow workflow) {
-    final PreesmScenario scenario = (PreesmScenario) inputs.get(AbstractWorkflowNodeImplementation.KEY_SCENARIO);
     final PiGraph graph = (PiGraph) inputs.get(AbstractWorkflowNodeImplementation.KEY_PI_GRAPH);
-
-    final StaticPiMMFlatPiMMLauncher launcher = new StaticPiMMFlatPiMMLauncher(scenario, graph);
-
-    PiGraph result = null;
-    final Logger logger = PreesmLogger.getLogger();
-    VisitorOutput.setLogger(logger);
     logger.log(Level.INFO, "Computing Repetition Vector for graph [" + graph.getName() + "]");
     // Flatten the graph
-    result = launcher.launch();
+    final PiGraph result = PiSDFFlattener.flatten(graph);
 
     final Map<String, Object> output = new LinkedHashMap<>();
     output.put(AbstractWorkflowNodeImplementation.KEY_PI_GRAPH, result);